+++ changelog.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [Unreleased]
+
+### Added
+
+- Rename component feature: 
+  - Add `getReferenceValue`, `setReferenceValue`, `getLinkValue`, `addLink`, `removeLink`  and `replaceLink` inside `ComponentAttribute.js`.
+  - Add `setId`, `createAttribute` and `getAttributesByType` inside `Component.js`.
+  - Add `renameComponentId` inside `DefaultData.js`.
+
 ## [0.18.0] - 2023/07/20
 
 ### Added